Update Install Instructions
Debian/Ubuntu 19.04 I have updated the autotools installation commands below Debian as the old package was missing or obsolete.  I'll be attaching screen shots in my pull request.  Also, this is my first Github contribution ever.  Feedback regarding content, protocol, etc is very much welcome.  

Also added to the comments on lines 35 and 37 to include the sudo command for ubuntu/debian users.  Just to make it tidy for the user, so they do not have to hunt around if they are perhaps a novice and unfamiliar with the permission denied message they would be getting.